move Panel type in types
